[MODIFY]统一修改为公共字段

+++ common/common-upms/src/main/java/com/gitee/common/upms/dto/UserInfoDTO.java
@@ -45,5 +45,5 @@ public class UserInfoDTO implements Serializable {
 	 * 角色集合
 	 */
 	@ApiModelProperty(value = "角色标识集合")
-	private Integer[] roles;
+	private Long[] roles;
 }
